Syntax fixes

+++ ssgan.py
@@ -255,8 +255,8 @@ fixed_noise = torch.FloatTensor(batch_size, nz, 1, 1).normal_(0, 1)
 fixed_noise = _to_var(fixed_noise)
 
 
-d_gan_labels_real = torch.FloatTensor(batch_size)
-d_gan_labels_fake = torch.FloatTensor(batch_size)
+d_gan_labels_real = torch.LongTensor(batch_size)
+d_gan_labels_fake = torch.LongTensor(batch_size)
 
 real_label = 1
 fake_label = 0
